BitterBot_Unified_AGI_Platform/requirements.txt
# Core Dependencies
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
httpx==0.25.1

# Database
asyncpg==0.29.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Caching and Queuing
redis==5.0.1
aioredis==2.0.1
celery==5.3.4
kombu==5.3.4
pika==1.3.2

# AI/ML Frameworks
torch==2.1.0
torchvision==0.16.0
transformers==4.35.2
ray[rllib]==2.8.0
numpy==1.26.2
scipy==1.11.4
scikit-learn==1.3.2
pandas==2.1.3

# Vector Databases
faiss-cpu==1.7.4
chromadb==0.4.18
langchain==0.0.340

# Distributed Computing
dask[distributed]==2023.12.0
mpi4py==3.1.5

# Monitoring and Logging
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
structlog==23.2.0
python-json-logger==2.0.7

# Storage
boto3==1.29.7
minio==7.2.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilities
click==8.1.7
rich==13.7.0
tqdm==4.66.1
tenacity==8.2.3
humanize==4.9.0

# Testing (included in main for development)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# Development Tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.5.0